+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5 (Pending):
+* Add SessionManager.getSessionIDs() to expose the list of sessions in manager.
 * Add SessionInfo.getAttributeKeys() to expose the list of attributes in session.
 
 ## 0.4: